Refactor addDragAndDrop in playerRenderer
This is an important discovery. Previously, when a ship was dropped
on top of a pre-existing one, the coordinates of that position where
not accessed as the overlaying ship was preventing access to the cell.
Even if this scenario happens, we are still interested in getting the
coordinates for to call moveToClosestValidPosition with. The way this
has been solved is by accessing the second element in the stacking order
with the built inmethod elementsFromPoint() when the dropped position
contains a ship, which in this case, will be the cell we are interested
in. To make this ocurr only on overlapping ships, an if statement
that triggers when the cell is non existing is used