Fix bug in switchOrientation
When using moveToClosestValidPosition, if the function couldn't find a valid position before it would just accept it and place the ship. Which is totally forbidden. This has been fixed by the function returning a boolean that indeicates if the move has been successful, if it hasnt't, switchOrientation just restores the initial position and orientation of the Ship, effectively doing nothing and preventing an illegal placement