Fix problem with renderAttacks
Due to the new way of expressing the coordinates, the board object and the html board where not synced, due to the way that renderAttacks was set up. This has been fixed by applying a transormation of coordinates that correctly relates each board array item with each corresponding div. This does make the code a little bit longer and less readable, so I will be thinking on alternative approaches to solve this, however the priority right now is to make the app work and finish it. First solve, then refactor